refactor(output): split OutputHandler into modular components
- Extract LatexOutputUtils.ts: standalone LaTeX formatting utilities
- Extract OutputFileProcessor.ts: file processing and XML summary logic
- Simplify OutputHandler.ts to orchestration role
- Remove unused indentLatexFile(s) from IOutputHandler interface

No functional changes - code is now more modular and focused.
